Move irf_map.py to core.py
