+++ gammapy/irf/core.py
@@ -200,8 +200,10 @@ class IRFMap:
             Sliced irf map object.
         """
         irf_map = self._irf_map.slice_by_idx(slices=slices)
+
         if "energy_true" in slices:
             exposure_map = self.exposure_map.slice_by_idx(slices=slices)
         else:
             exposure_map = self.exposure_map
+
         return self.__class__(irf_map, exposure_map=exposure_map)
